fix: logout

+++ server/dao.js
@@ -146,11 +146,11 @@ exports.generateMeme = (meme) => {
   });
 };
 
-// delete an existing task
-exports.deleteMeme = (user, id) => {
+// delete an existing meme by user who created the meme
+exports.deleteMeme = (userId, id) => {
   return new Promise((resolve, reject) => {
-    const sql = "DELETE FROM memeList WHERE id = ? ";
-    db.run(sql, [id], (err) => {
+    const sql = "DELETE FROM memeList WHERE id = ? and userId = ? ";
+    db.run(sql, [id, userId], (err) => {
       if (err) {
         reject(err);
         return;
